work: refactor so that the module can load without throwing on other operating systems and runtimes
